--- src/main/kotlin/AppMenuSelection.kt
@@ -1,4 +1,0 @@
-interface AppMenuSelection {
-    fun showMenu()
-
-}
