--- src/main/kotlin/Note.kt
@@ -1,3 +1,0 @@
-import java.util.Scanner
-
-class Note(val noteName: String, val noteData: String)
